Ensure placement requests are withdrawn
Looking through the work we’ve done, I noticed that there was a check
in withdrawPlacementRequestWhenPlacementApplicationWithdrawn that meant
only workflow managers could withdraw placement requests associated with
the placement application, when actually we want the creators of the
placement application (who won’t necessarily have that role) to be
able to withdraw the associated placement requests too.

I’ve moved the relevant code into the placement application service, as
well as adding a few unit tests to ensure permissions are respected and
emails are sent correctly.