Buffer `WebClient` responses to avoid `DataBufferLimitException`
The memory limit for responses from the Delius Integration API has
recently been increased to 750,000 bytes (see #1120). However, this was
still not sufficient when generating booking reports, as personal
information is requested for several hundred CRNs concurrently.
Continuing to increase this memory limit can only achieve so much as the
API is limited by how much memory it has access to. This introduces a
buffering approach for the responses which circumvents the limit and
also avoids constructing potentially large strings as an intermediate
step.