Refactor SubmittedApplication transformation to dedicated class
We'd need to add another transformer to the class to handle
the representation of StatusUpdates, in the case of
SubmittedApplications only.

Rather than adding more complexity, let's use a dedicated
SubmittedApplication transformer which can deal with the
needs of the SubmittedApplication as distinct from the "in-
progress" pre-submission use cases.

Immediately we remove the need for the NomisUser
transformer to be in the ApplicationsTransformer.