Use explicit query for getting placement requests
The findAllByAllocatedToUserIdAndReallocatedAtNullAndIsWithdrawnFalse method in particular was involved in the investigation of recent out of memory error [1,2].

The code is getting all placement requests and then putting it through a transformer loop which then goes and queries for each association: applications, placement_applications, assessments and placement_requirements - N+4.[3]

To help with performance we join this data together up front so that requets to eg. `it.application` don’t require another query.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-ui/blob/main/server/%40types/shared/models/PlacementRequest.ts
[2] https://mojdt.slack.com/archives/C044ZUGJRC7/p1699440674447369
[3] https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/3ae3df2f57dc682a8fef202daa0c11e11761f7ab/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/controller/PlacementRequestsController.kt#L65