Omit placement application data column from placement request query
We don’t need the data blob so can save on performance by filtering it out.

To do this we have to be explicit with the columns we care about, rather than getting all with *. This is arguably a safer place to be for performance but requires any new columns to be copied in.