Use streaming correctly when using large dataset
When looking at https://www.baeldung.com/spring-data-jpa-iterate-large-result-sets#streaming-from-the-database,
I missed that the `try` block used the output of `findAllForService`
within an argument for the data to stream correctly. This way to do
this in Kotlin is with a `.use` block, so I’ve updated the code to do
this.