Only update applications without statuses
We’re still having problems in preprod when updating and pages getting
skipped. To get around this, I’ve changed the query to only fetch
applications without a status. This borrows from an approach we used in
https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/906
when we were updating offender names, which seemed to work.

Only fetching applications without a status also means we can run the
task multiple times and only fetch/update a subset of applications too.