Detekt monitors for wildcard java util imports
This keeps the rules inline with how we’ve been importing on the project so far. With this change we don’t review whether this is the right/wrong choice, we just aim to consistently apply the previous rule.

Detekt defaults this to be excluded[1] so we remove the exclusion.

[1] https://detekt.dev/docs/rules/style/#wildcardimport