Summary returns ApprovedPremisesApplicationStatus
Before, the application statuses were based on a number of
different SQL subqueries, which was unweildy to maintain, and
made querying by status tricky. We now have a status database
field, which gets updated on the fly. This removes a lot of
the subqueries from our summary queries, as well as updating the
API to return an ApprovedPremisesApplicationStatus, based
on the status field stored in the database.