Stub out call to Manage-Users /externalusers/{username} endpoint
We continue to verify that a previously unseen assessor has
an "external user" record persisted when viewing a submitted
application (of list of submitted applications) for the
first time.

Rather than building the record from information in the JWT,
we use the response from the Manage-Users `/externalusers/{username}`
endpoint.

NB: it's essential to register the url of a new API in
`application-test.yml` e.g.

```
base-url: http://localhost:#WIREMOCK_PORT
```

else the wiremocks will not be picked up and you will be
scratching your head...