Placement with a cancelled booking can be rebooked
This adds a `isCancelled` helper to a booking to check if a booking is
cancelled, we can then use it to add to the
`createApprovedPremisesBookingFromPlacementRequest` validations to
allow placement requests with cancelled bookings to have a new booking
applied to them.

I’ve also DRYed up the expectations in the unit tests a little bit.
There’s still a bit too much repetition in the setup stage than I’d
like, but it’s a bit better than it was.