Add a job to update the sentence type/situation
This adds a new data migration job to backfill the sentence type and
situation. To keep things contained I’ve created a new repository and
put the helper classes in the job itself.

As we’re only fetching a small subset of data, we _should_ be able to
get away without having to paginate through the results, but the proof
of the pudding will be in the eating when we run it in preprod!