Migration for person details to cas2 applications
Currently, we store the CRN of a person on an application, and then use
that CRN to make an API call to retrieve the rest of the person's data
whenever we return an application or application summary.

Instead, we would like to store a "snapshot" of the information about
the Person on an CAS2 application, at the moment when the application is
created.

We would like to save these fields on an Application Entity, which will
be used to build the Person field on an Application that is returned to
the frontend.