Add AutoScript 'enabled' config
AutoScripting is a form of seeding which does not use CSV
seeding files. i.e. the seeding is entirely defined in a
script. In CAS2 we will use this for seeding using real
NomisUsers and ExternalUser (in dev and test environments)
whose usernames we don't want expose in a seed file.