Add a BookingSummaryForAvailability interface
This will be used in a custom query so we can get whether a group of
bookings is arrived, not arrived or cancelled in one query, rather than
having to fetch each entity for every single booking.

I’ve also added a Factory for use in the unit tests. This adds a
`BookingSummaryForAvailabilityImpl` class that implements our
interface, then uses the usual factory pattern we use elsewhere for
“normal” JPA entities. I’ve had to name the initializer vars with the
suffix `Var` as we get a naming clash because Kotlin attempts to create
getter funcions with the prefix `get` for each variable, which overrides
the behaviour we set in the class body and causes the compiler to blow
up