Update findAllByPremisesIdAndOverlappingDate
This changes the findAllByPremisesIdAndOverlappingDate query to count
the cancellation, nonArrival and arrival entities, rather than loading
them in on each booking. This is then returned as a
`BookingSummaryForAvailability` DTO using a JPA Projection.

I’ve also added another integration test to ensure the availability
response is correct too, as we were only unit testing it previously.