Fix cancelledAt time in cancellation Domain Event
When converting the cancelledAt LocalDate into an Instant for the Domain Event we were using an incorrect Offset Time. This led to a Date Time that would always be set to the previous day. The date conversion has been changed to explicitly set the time to midnight to avoid this issue.

Coverage has been improved in the corresponding Integration Test to check the Domain Event's cancelledAt date/time (something which appears to of been accidentally lost in a previous refactor)