Added a new column to the approved premises application report showing when assessment was last allocated
Whilst we are creating multiple assessments for one application in the test (by reallocating the initial assessment), we have no way of asserting that the allocation date is coming from the reallocated assessment entry because the date is the same on both assessments and there is no time element in the report.

If we could control the ‘clock’ used by the application during tests we would be able to test this, but such a change is non trivial.

---

Some refactoring of the existing report tests was required to support testing the additional report column and to keep detekt happy. Specifically:

* Introducing an ApplicationFacets data class to encapsulate parameters previously defined on the 'assertApplicationRowHasCorrectData' function, ensuring the parameter count remained <= 6
* Removing some duplication of logic used to find an application's report row, and moving that logic into 'assertApplicationRowHasCorrectData', ensuring calls to the method don't exceed detekts max line length

These changes also help prepare for adding/testing additional report columns