Use allocations engine to fetch users for a Task
When returning the eligible users for a Task, we’ve had issues where
ineligible users are being returned. Tracking this down, it seems to be
a problem where the `hasQualificationsAndRoles` specification is
returning users that have any qualification in the provided list of
required  qualifications, not all of the roles.

This is a situation we’ve already covered when automatically allocating
users to a task, so I’ve updated the UserAllocationsEngine (which
handles automatic allocations) to accept an `excludeAutoAllocations`
flag. This allows us to return all qualfied users in our pool of
potential users (currentlt the AllocationsEngine excludes users who
have roles that prevent them from being automatically allocated a task)

I’ve also made some tweaks to the Tasks tests to make sure the
expected users have the correct roles.