Refactor DomainEventService tests to use Nested structure
There is no change to the tests, this commit is purely
structural, using the @Nested annotation to group the
existing specs which describe the persistence and retrieval
of CAS2's ApplicationSubmitted domain event:

- by function name
  - by scenario

The overall spec now reads:

```
DomainEventServiceTest:
  ApplicationSubmitted:
    GetCas2ApplicationSubmittedDomainEvent:
      - returns event
      - returns null when event does not exist
    SaveCas2ApplicationSubmittedDomainEvent:
      - persists event, emits event to SNS
      - does not emit event to SNS if event fails to persist to database
      - does not emit if emitDomainEventsEnabled is false
```

This is a preliminary step before adding similar specs
around the new CAS2 ApplicationStatusUpdate domain event. It
should lead to a spec file which, though long, remains
readable and manageable.