Seed a premises for jimsnow for local e2e tests
As jimsnowldap is the user CAS3 sign in with and run our e2e tests with locally we want at least one premises to be in place before we run the e2e tests.

Currently we are seeing failures early on in some e2e tests[1] where they look to find an active premises but none exists. I’m not clear why this has only recently started to fail. Perhaps Cypress was running in a set order and the create premises test was putting one there for us?

We do know that that ‘create premises’ e2e test is covering the creation process by creating a new record. By seeding another premises we shouldn’t be reducing our coverage.

We seed in both dev+test and local so the premises are consistent between and you can change which API the e2e tests use.

[1] https://app.circleci.com/pipelines/github/ministryofjustice/hmpps-temporary-accommodation-ui/4808/workflows/ee83177b-116a-4401-9246-22ec233cf8da/jobs/10880