`getDateCapacities` only looks one year ahead
In prod, we’re noticing a lot of timeouts in prod/preprod owing to the
fact that some bookings have been made with an end date of 2924, meaning
we loop through thousands of days, causing a timeout. Rather than adding
validation for non-silly periods (there may be a valid reason for a user
to block a bed, rather than it being a typo), we ensure that the longest
the `getDateCapacities` function looks ahead is 1 year. This should
help the timeout issue, as well as giving us a reasonable amount of
capacity information