Refactor seed fixture fixture loading: as 'resource'
Loading these fixture files with `FileUtils.readFileToString()`
does not work in our deployed environments on the Cloud
Platform where artefacts are bundled into a JAR artefact.

In order to load a file from the 'resources' bundle we need
to use some special functions: we use a special "resource
loader" to generate a stream which we then pass to
`FileCopyUtils.copyToString()`

See https://www.baeldung.com/spring-load-resource-as-string

We wrap the resource loading code in a `try` block
so that the seeding migration process will continue in event
of problems. In this way this "auto-script" seeding should
no longer cause a deployment failure.