Start emitting 5 CAS3 domain events
We’ll come back and do the remaining 4 in the new year as guidance and comms need to take place with our users first.

As we won’t be emitting arrival and departed events for a few weeks there will be a gap where some bookings only get as far as confirmed.