Turn off all domain events in dev
We are trying to add load tests as part of CI via Gatling. As these will send many, many thousands of requests we should prevent this level of traffic being consumed by integrations and all that data going into ndelius etc.

This level of data may affect their performance and/or produce a huge amount of data that we and other services are then unable to load.

Domain events should still be enabled in all other environments and can be tested by the integrations and ndelius team in pre-prod.