Refactor `TasksController` to use closures to retrieve person name
Currently the call to `OffenderService.getOffenderByCrn` is being
performed in such a way that it's not easily aggregated. By extracting
the logic used to retrieve the person name into helper functions and
passing in closures that call the functions, it preserves the existing
behaviour while making the transition to aggregate methods for the
list endpoints tractable.