Use aggregate method in TasksController
This gathers up all the CRNs from the task(s) in the controller and
passes them to the new `getOffenderSummariesByCrns` method. These
resulting summaries are then passes into the `getPersonNameFrom*`
functions to allow up to find a summary for the CRN and then return the
name (via the `getNameFromPersonSummaryInfoResult` helper function.