Simplify DomainEvent-related classes
After approving #1224, I noticed that we didn’t need to get the
domainTopic in both the service and the worker. This simplifies the
calls to the worker by removing the `hmppsQueueService` from the Service
and fetching the `domainTopic` when we initialize the
`ConfiguredDomainEventWorker`. We can then set up the `publishEvent` in
the worker classes, rather than passing it from service to worker.

I’ve also made some tweaks to the tests, using `mockkConstructor` to
tighten up the expectations in the `ConfiguredDomainEventWorker` tests,
and making the mock call expectations tighter using a
`matchingPublishRequest` function.

I also noticed that the verifications in the `DomainEventServiceTest`
were asserting that `mockHmppsTopic.snsClient.publish` was not being
called, which would always succeed. Now we change those expectations to
be on the worker class method - `domainEventWorkerMock.emitEvent`