Enable 'auto-script' seeding in local development environment
This ought to have been enabled. In the local development
environment the `Cas2AutoScript.script()` process is used to
create seed fixtures for CAS2. At present these comprise a
number of applications and status updates in various states.

There are no API calls or known downsides to having this
enabled.