pin gatling to a specific version
By default the gradle gatling plugin will automatically chose with version of gatling to use. Gatling >= 3.10 is incompatible with the version of Jackson used by Spring Webflux. This causes a ClassNotFoundException when running load tests.

The workaround applied on this commit is to pin the version of gatling used by the plugin to a version < 3.10.

We can consider unpinning the version once we have upgraded to a later version of webflux (and even, this is assuming at some point spring boot 3 will use jackson >= 2.16.0, see notes below)

Specific details on the incompatibility:

* Gatling >= 3.10 uses jackson 2.16.0
* Earlier versions of Gatling use jackson 2.15.0
* It appears that Gatling is using jackson 2.16.0 in such a way that will fail if 2.15.0 is used at runtime
* We currently use jackson 2.13.5 at runtime, provided as a transitive dependency from spring webflux
* Even if we try to force spring webflux to it’s latest version, it still only uses jackson 2.15.0
* If there is ever a version of spring webflux that uses jackson >= 2.16.0 we can consider unpinning the gatling version