Add a TaskEntity with a query to fetch all tasks
This adds a new `TaskEntity` which is not tied to any particular
table, we then have a `UNION` query to fetch all “Taskable” entities
in one database call. Because we have to specify a `countQuery` (and to
save repeating ourselves), I’ve added a `ALLOCATABLE_QUERY` constant
that we can reuse when counting the results for pagination.