Add a `getAllReallocatable` service method
This calls the `getAllReallocatable` repository method to get
all the entities which can be tasks. We then grab all the IDs and
carry out a query on the relevant repository. I’ve also added a
`TypedTask` sealed class which allows us to return a list of multiple
types, but still allow stricter typing than just returning a `List<Any>`