Use the new service method in the controller
Now we can fetch all the tasks and carry out native pagination, we fetch
the CRNs from the resulting entities and transform them to the API
types. I’ve done a small bit of refactoring on some of the tests to make
them easier to follow, but the expecations are largely the same.