Define /cas2/reports/{reportName} endpoint
Reports will have human readable names, e.g.

- `submitted-applications` for a list of `application.submitted`
  domain events

- `application-status-updates` for a list of
  `application.status-updated` domain events

Note that we plan to make the CAS2 reports much dumber
than those of CAS1 and CAS3. We intend to expose sufficient
information from our domain events and from
submitted applications to allow analysts build their own
reports. We expect that they will retrieve or "join" linked
data from elsewhere using tools and platforms such as:

- Performance Hub
- Analytic Platform
- Power BI
- Tableau

We don't plan to:

- retrieve or store data which might be desired soley by
consumers of reports, if that data isn't needed to run the
CAS2 digital service

- perform aggregations, analytics or processing of data.
These reports are intended to be raw material for data
analysts

- supply operational reports. These reports supply metrics
for measuring the performance and usage of the service