Describe behaviour when referringPrisonCode can't be retrieved for domain event
The call to the Prison API to find the referringPrisonCode when saving
the `application.submitted` domain event *should* never 404 or otherwise fail,
as when creating  the application initially a similar call was made.

If there is a problem with accessing the Prison API, we fail hard and
abort our attempt to submit the application.