Add 'first class' properties to CAS2 Application
These are properties which have been identified as important
metrics to be exposed in both:

- domain event (`application.submitted`)
- downloadable report ("Submitted applications")

It's important to "promote" them out of the
`application.data` JSON blob so that they can be found in a
consistent place e.g. for use-cases such as:

- inclusion of "preferred area" as meta data in a dashboard
- building a list in the service of submitted application,
ordered by conditional release date

NB: the location of data within the `application.data` blob is
inherently unstable by design as we expect the format of the
application questionnaire to change over time.