Set additional application.data properties sent by UI
These three additional properties which are needed on the
`application.submitted` domain event and on the "Submitted
applications" download originate from the questionnaire
represented by the `application.data` JSON blob. They are
being promoted to first-class attributes of the application
for resilience, as the structure of the questionnaire is
expected to change over time, but we need to be able to
reliably use these data as meta-data e.g.

- to display in the summary of a submitted application
- to show in a dashboard listing applications
- in a query to select a subset of applications

As the UI (the front-end app) is responsible for the
changing structure of the questionnaire over time, it makes
sense for the properties to be pulled out of `application.data`
by the UI, at the time of submission (POST /submissions) as
distinct params. This is the approach already in place in
other CAS services, see:

https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/af10cdc47422a157fb775443b55083f4eed4ee5f/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/ApplicationService.kt#L844-L856

Therefore we add 3 new properties to the OpenAPI `SubmitCas2Application`
schema.