Add utility methods to convert between `CaseAccess` and `UserOffenderAccess`
These methods are useful for transforming the results from the AP Delius
Context API for user access into those from the Community API, and vice
versa.

Converting from `CaseAccess` to `UserOffenderAccess` is lossy as the
latter does not contain fields to represent the exclusion message or the
CRN. However, the inverse conversion is lossless.