Add `map` and `flatMap` utility methods to `ClientResult`
`ClientResult.map` transforms the body of a `Success` variant according
to the provided transformation function, e.g.:

```
val result = ClientResult.Success(HttpStatus.OK, 1, false)

result.map { it * 2 } // ClientResult.Success(HttpStatus.OK, 2, false)
```

For the `Failure` variants, the `ClientResult` is cast to a result of
the target type without transformation. This is safe as these variants
contain nothing of the original response type.

`ClientResult.flatMap` transforms the body of a `Success` variant into
multiple results according to the transformation function. Each result
is wrapped in a `Success` containing the same metadata as the original
`Success` value. For the `Failure` variants, the `ClientResult` is cast
to a result of the target type without transformation and wrapped in a
singleton list. This is also safe for the same reason as above.