Dev user seed doesn’t conflict on ID
Id is a randomly generated uuid so this will never conflict.
Removing the parameter then allows any type of conflict to result in DO NOTHING.
This allows us to then add a new database uniqueness constraint to fix this across the board, since arguably this scenario should not be possible.