Return a count of all tasks
Before we were fetching only the Assessments, but we should
be returning a count of all a user’s tasks ideally, so I’ve updated
the query to return placment applications and requests, and then
summing them all up in the service. I also noticed a bug in
`Given a Placement Request` where we were allocating the associated
assessment with the `placementRequestAllocatedTo` user, rather than
the `assessmentAllocatedTo` user, which was not getting used anywhere!

I’ve also updated the task count test to be (hopefully) a bit more
readable now we’re dealing with a bunch more new task types