add environment variable for cas2 assessors inbox
When a CAS2 application is submitted, we need to send an email to
the CAS2 assessors with a link to the application so they can assess it.

In production, this will be an email address for the third party
organisation doing our assessing.

In all other environments, we will use a test email address that our dev
team have access to, so that we can ensure the integration is working.

We are proposing to introduce this email value as a namespace secret, so
that we can change it based on the environment/if our third party email
address changes, without having to re-release the API code. Also so that
we don't commit potentially sensitive email addresses to the github
repo.