Add pagination to `/assessments`
This adds support for pagination to the `/assessments` endpoint. For
speed of implementation, we’re just doing the pagination by way of
splitting the resulting list, so we can get on with implementing the
frontend. We can come back and do the database-side pagination later.

I’ve also made a few tweaks to the tests to seperate them out and
DRY up some of the expectations.