Seed CAS2 application.submitted domain events
For each CAS2 application which is seeded (in a submitted
state) we create an corresponding `application.submitted`
domain event.

We make `ApplicationService.createCas2ApplicationSubmittedEvent()`
public to make this easier. Having a public service method
will be helpful in the future when, no doubt, we need to
backfill or replace domain events due to operational
problems.