Seed CAS2 application.status-updated domain events
For each CAS2 status updates which is seeded we create a
corresponding `application.status-updated` domain event.

As in the preceeding commit, we make
`StatusUpdateService.createStatusUpdatedDomainEvent()`
public to make this easier. Having a public service method
will be helpful in the future when, no doubt, we need to
backfill or replace domain events due to operational
problems.