Move CAS1 Assessment Status filtering into SQL
This change is a pre-requisite to enable paging/sorting in SQL instead of programmatically in the controller code

Not only is the status filtered in SQL, but the SQL now also returns a Status String, meaning we no longer have to determine it programatically when listing assessments.

Unfortunately, we still determine the status programmatically when retrieving an individual assessment (in AssessmentTransformer), so we have a duplication of this logic in both Java and SQL. The presumed way to address this issue is to explicitly store the status in the assessment table itself, which then allow us to simplify the SQL added by this commit.

Note that we no longer return dateOfInfoRequest and isStarted from the assessment search queries because this was only ever used to determine the CAS1 status programatically for search results

I’ve temporarily suppressed a CyclomaticComplexMethod issue in the AssessmentsController as this method will be refactored in a subsequent commit as part of the work to move CAS1 paging into SQL