Use `equals` to match `HttpStatus`
We’ve been having an issue when a user hits the `userAccess` endpoint,
but gets an error. `canAccessOffender` should be catching when the
endpoint returns a 403, but the conditional is not working as expected.

After some research, I’ve discovered that `HttpStatus` is an enum, so
`==` is using reference equality, not value equality. However, using
`equals` uses the value, so will match predictably [1]

[1]: https://stackoverflow.com/questions/45975550/using-equal-sign-for-httpstatus-bad-request-not-working