Permissions for CAS2 users based on ROLE_POM
Previously we were granting access to all users with ROLE_PRISON, a role
granted automatically to all nomis authorised users. This is too broad a
restriction, because it means any users logging in via nomis will be
able to search for prisoners and make applications.

As our referrer users for private beta will be POMs, we should restrict
to the POM role for now. This can be assigned to devs to test in
different environments.

Contract managers will be given the role CAS2_ADMIN so they can see
submitted applications, but they shouldn't need ability to create
applications.