Use DB-based paging for CAS3 Assessments
As part of this change a default sort field is now assumed if one is not provided by the API call. This simplifies handling of the paging criteria and ensures we have consistent results from paging.

Unused methods have been removed from paging utils, and the default page size is now taken from config instead of hard-coded