Create rule-based allocation engine
The `UserAllocator` class can select a suitable user to allocate to one
of an assessment, placement request, or placement allocation, according
to a configurable list of rules.

A rule must implement the `UserAllocatorRule` interface. The interface
provides the `evaluateAssessment`, `evaluatePlacementRequest`, and
`evaluatePlacementApplication` methods, which can be overridden to
implement the rule logic. Rules can be ordered by defining their
`priority`; a rule with a lower value will be evaluated earlier.

When a rule is evaluated, it will return a `UserAllocatorRuleOutcome`
representing what action the allocator should take. Currently the
available outcomes are:
- `Skip`: The default rule outcome. This indicates that the allocator
  should proceed to the next rule.
- `DoNotAllocate`: This indicates that the allocator should abort and
  not allocate a user.
- `AllocateToUser`: This indicates that the allocator should finish
  evaluating rules and allocate to the given user. If this allocation
  fails, however, a warning is logged, and the rule is skipped.