Exit remote_db if downloaded jaqy fails checksum
If the downloaded version of jaql fails the checksum we have
committed to the repo:

```
sha256sum --check jaqy-1.2.0.jar.checksum.sha256
```

then we exit the `remote_db` script and instruct the user to
investigate.

There's a risk that the checksum failure could be due to
the downloaded file (a "release" on GitHub) having been
maliciously modified.