Add configuration for legacy allocation rule
The `user-allocations.rules.legacy-allocator.enabled` and
`user-allocations.rules.legacy-allocator.priority` configuration values
will allow for the legacy `UserAllocationsEngine` to be used
per-environment through an adaptor rule. This will allow the new user
allocator to be introduced without any external change in behaviour, or
for a rollback to the legacy behaviour to be performed without reverting
code changes.