Improve CAS2 'Auto-script' seeding to include properties set on application submission
These properties are set on submission of the application in
`ApplicationService.submitApplication()`:

"promoted" up out of the application.data JSON blob:

- `preferredAreas`
- `hdcEligibilityDate`
- `conditionalReleaseDate`

and also `referringPrisonCode` which is looked up from the
Prison API.

Another property (`telephoneNumber`) will be added soon.

These first-class properties are used in the 'metadata'
which is shown in the "application summary" which
accompanies a submitted application. They are also included
in the submitted.application domain event and through this,
the "Submitted applications" management information report.

Note that the new `applyFirstClassProperties(application)`
uses `saveAndFlush()` from [JPA Repository] [JPA_Repository]
rather than `save()`. This is purely for the sake of the
specs as we want to continue to demonstrate that exactly 3
applications are seeded for each NOMIS user:

```kt
fun `creates 3 applications for each Nomis User`() {
  autoScript.script()

  verify(exactly = 3) { mockApplicationRepository.save(any()) }
}
```

[JPA_Repository]:
https://docs.spring.io/spring-data/jpa/docs/current/api/org/springframework/data/jpa/repository/JpaRepository.html