Fix time conversion test instability
There is a need to be able to remove the ‘nano-specific’ part of milliseconds in tests because this nano precision is lost when persisted to postgres, leading to dates not matching a copy of the same date retained in memory that still has nano-specific values.

The previous implementation of the method could lead to invalid nano values. This has been solved by instead using in-built TimeUnit conversion to deal with rounding.