Enable TEST_AND_GUEST_LIST mode in preprod
We would like to test the end-to-end journey in CAS2, which means we
need to trigger emails on application submission.
This sets the correct Notify Mode, to trigger emails only to Notify team members and the guest list.
Although we will only use the Guest List Client and API Key, the code was not able to
compile without also including the regular API Key.
I've created new keys for the preprod environment, and added them to the kubernetes namespace secrets.